requirements-doc.txt
autoflake
isort
black

# pytest
pyvirtualdisplay
discogui
easyprocess
pillow>=8.1.0
pygetwindow
# https://github.com/Kalmat/PyGetWindow/archive/refs/tags/v0.0.10-beta.zip
